removed commented and duplicated code, pulled out room creation into /widget path
