+++ lang/main.json
@@ -854,10 +854,8 @@
         "enterRoomTitle": "Start a new meeting",
         "footerText": "Requires ",
         "getHelp": "Get help",
-        "roomNameAllowedChars": "Meeting name should not contain any of these characters: ?, &, :, ', \", %, #.",
         "go": "Start video call",
         "goSmall": "Start video call",
-        "join": "CREATE / JOIN",
         "info": "Info",
         "join": "CREATE / JOIN",
         "moderatedMessage": "Or <a href=\"{{url}}\" rel=\"noopener noreferrer\" target=\"_blank\">book a meeting URL</a> in advance where you are the only moderator.",
